move Endpoint to a Trait
